MODUSERS-242: Upgrade to RAML Module Builder 32.x
Most changes are explained in the RMB upgrade guide:

github.com/folio-org/raml-module-builder/blob/master/doc/upgrading.md

In addition, tenant setup seems to have changed from a synchronous to an
asynchronous operation.  Tests in this module that required a tenant to
operate were written with synchronous setup in mind-thus, when the
upgrade was applied, they all started failing.  All tenant initialization
code had to be re-written to allow for asynchronous operation, which accounts
for most of the changes made to tests.

The tenant ref API also had to be reworked-my understanding from the
RMB folks is that overriding loadData has been the preferred method of
loading reference and sample data to test tenants for a while now.

Refs: https://issues.folio.org/browse/MODUSERS-242