MODUSERS-260, MODUSERS-261: Remove jsoup HTML stripping
Remove jsoup HTML stripping.

It is not needed because Stripes/React can handle
special characters and prevent HTML injection.

Stripping characters must be declared in the RAML schema
https://github.com/folio-org/mod-users/blob/master/ramls/department.json
after the User Management SIG and the PO have agreed.

If some characters should be removed or replaced
the backend must do it consistently. This is not the case,
see examples on https://issues.folio.org/browse/MODUSERS-260

The used version of jsoup has a security vulnerability:
https://cve.mitre.org/cgi-bin/cvename.cgi?name=CVE-2021-37714